Add files via upload
The dataset was cleaned using Power BI to ensure data quality by checking for duplicates and empty spaces. Additionally, the first names and last names columns were merged into a single column labeled "Names," and the first row was utilized as the header for the dataset. A relationship structure was established within the dataset, which comprises four tables, each representing different aspects of gym data. These relationships facilitate cohesive analysis across the tables. A dashboard was developed to visualize gym performance metrics and other analyses. This provides a comprehensive overview of the data insights.